requirements.txt
torch==2.3.0
torchvision==0.18.0

pretrainedmodels==0.7.4
efficientnet-pytorch==0.7.1
timm==1.0.3

tensorboard==2.16.2
opencv-python==4.10.0.82

pandas==2.2.2
path==16.14.0
easydict==1.13
scipy==1.13.1
matplotlib==3.9.0
numpy==1.26.4
tqdm==4.66.4
PyYAML==6.0.1
pillow==10.3.0
scikit-learn==1.5.1

h5py==3.12.1
librosa==0.10.2.post1

# lightgbm==4.5.0
